server/requirements.txt
# AI Financial Advisor - Python依赖包
# 更新日期：2025-10-23

# ============================================================================
#                           核心框架
# ============================================================================

# FastAPI - 现代化的Web框架
fastapi==0.104.1

# Uvicorn - ASGI服务器
uvicorn[standard]==0.24.0

# Pydantic - 数据验证
pydantic==2.5.0


# ============================================================================
#                           认证和安全
# ============================================================================

# Passlib - 密码加密
passlib[bcrypt]==1.7.4

# Python-JOSE - JWT token处理
python-jose[cryptography]==3.3.0

# Python-Multipart - 表单数据处理
python-multipart==0.0.6


# ============================================================================
#                           HTTP请求
# ============================================================================

# Requests - HTTP客户端
requests==2.31.0


# ============================================================================
#                           开发工具（可选）
# ============================================================================

# Python-dotenv - 环境变量管理
python-dotenv==1.0.0


# ============================================================================
#                           说明
# ============================================================================

# 安装方法：
#   pip install -r requirements.txt
#
# 开发环境安装（包含开发工具）：
#   pip install -r requirements.txt
#
# 生产环境部署建议：
#   - 使用虚拟环境：python -m venv venv
#   - 激活虚拟环境：
#       Windows: venv\Scripts\activate
#       Linux/Mac: source venv/bin/activate
#   - 安装依赖：pip install -r requirements.txt
#
# 版本说明：
#   - 使用 == 固定版本号，确保部署一致性
#   - 定期更新依赖包以获取安全修复
#   - 升级前在测试环境验证兼容性

